Reorganize to break out common helper function.
